requirements.txt
algoliasearch==2.6.3
algoliasearch-django==2.0.0
APScheduler==3.10.4
asgiref==3.8.1
asttokens==2.4.1
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.1.0
cloudinary==1.39.0
colorama==0.4.6
comm==0.2.0
cryptography==41.0.2
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
dj-database-url==2.1.0
Django==5.0.6
django-allauth==0.60.1
django-apscheduler==0.6.2
django-background-tasks==1.2.5
django-cloudinary-storage==0.3.0
django-compat==1.0.15
django-cors-headers==3.14.0
django-crontab==0.7.1
django-environ==0.11.2
django-filter==23.2
django-jazzmin==3.0.0
django-ranged-response==0.2.0
django-resized==1.0.2
django-simple-captcha==0.6.0
django-templated-mail==1.1.1
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
drf-nested-routers==0.93.4
drf-spectacular==0.27.2
dronekit==2.9.2
dronekit-sitl==3.3.0
environ==1.0
executing==2.0.1
filelock==3.10.1
future==0.18.3
greenlet==3.0.3
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
ipykernel==6.27.1
ipython==8.18.1
jedi==0.19.1
jsonschema==4.17.3
jupyter_client==8.6.0
jupyter_core==5.5.0
lxml==4.9.2
matplotlib-inline==0.1.6
MAVProxy==1.8.59
monotonic==1.6
nest-asyncio==1.5.8
numpy==1.24.2
oauthlib==3.2.2
packaging==23.2
parso==0.8.3
Pillow==9.5.0
platformdirs==3.1.1
prompt-toolkit==3.0.42
psutil==5.9.4
psycopg2==2.9.9
psycopg2-binary==2.9.9
pure-eval==0.2.2
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
pymavlink==2.4.37
pyproject-toml==0.0.10
pyrsistent==0.19.3
pyserial==3.5
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
pywin32==306
PyYAML==6.0
pyzmq==25.1.2
requests==2.30.0
requests-oauthlib==1.3.1
shortuuid==1.0.13
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
SQLAlchemy==2.0.30
sqlparse==0.4.3
stack-data==0.6.3
toml==0.10.2
tornado==6.4
traitlets==5.14.0
typing_extensions==4.10.0
tzdata==2023.3
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.0.2
virtualenv==20.21.0
wcwidth==0.2.6
whitenoise==6.6.0
